renamed projects folder and files to project to match command name change
