+++ packages/cli/bin/cli.js
@@ -28,7 +28,7 @@ const functionsCommand = require('../commands/functions');
 const listCommand = require('../commands/list');
 const openCommand = require('../commands/open');
 const mvCommand = require('../commands/mv');
-const projectsCommand = require('../commands/projects');
+const projectCommands = require('../commands/project');
 const themeCommand = require('../commands/theme');
 const configCommand = require('../commands/config');
 const accountsCommand = require('../commands/accounts');
@@ -94,7 +94,7 @@ const argv = yargs
   })
   .command(openCommand)
   .command(mvCommand)
-  .command(projectsCommand)
+  .command(projectCommands)
   .command(themeCommand)
   .command(configCommand)
   .command(accountsCommand)
